src/backend/requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.50.0
greenlet==3.0.3
imageio==2.34.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.2.0
PyJWT==2.8.0
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
SQLAlchemy==2.0.29
threadpoolctl==3.4.0
tifffile==2024.2.12
typing_extensions==4.10.0
tzdata==2024.1
uWSGI==2.0.24
Werkzeug==3.0.1
